feat: MediaPipe SPM 연동 및 다트 포즈 분석 엔진(Python -> Swift) 포팅 완료
CocoaPods 충돌 해결, 모델 파일 번들 추가, SDK API 마이그레이션 및 SwiftUI 최신 문법 업데이트 적용

+++ DartPose/Models/Config.swift
@@ -33,9 +33,9 @@ enum DartConfig {
     static let validationMaxReleaseTimingMs: Double = 1200.0
     static let validationMinRomAngle: Double = 30.0
     static let elbowStabilityThreshold: Double = 0.005
-    static let takebackMinAngle: Int = 30
-    static let takebackMaxAngle: Int = 110
-    static let elbowExtensionVelMin: Int = 150
+    static let takebackMinAngle: Double = 30
+    static let takebackMaxAngle: Double = 110
+    static let elbowExtensionVelMin: Double = 150
     static let bodySwayThreshold: Double = 0.05
     static let shoulderStabilityThreshold: Double = 0.003
     static let throwMinFrames: Int = 15
